Move address encoding functions from Shelley.Compatibility to Address.Encoding.
Remove EpochNo types from Cluster module.
