+++ lib/local-cluster/local-cluster.cabal
@@ -61,6 +61,7 @@ library
     , cardano-wallet-application-extras
     , cardano-wallet-launcher
     , cardano-wallet-primitive
+    , cardano-wallet:address-derivation
     , cardano-wallet:cardano-wallet-api-http
     , cardano-wallet:cardano-wallet-local-cluster
     , contra-tracer
