+++ lib/wallet/integration/src/Test/Integration/Scenario/CLI/Byron/Addresses.hs
@@ -13,14 +13,14 @@ module Test.Integration.Scenario.CLI.Byron.Addresses
 
 import Prelude
 
+import Cardano.Wallet.Address.Encoding
+    ( encodeAddress )
 import Cardano.Wallet.Api.Types
     ( ApiAddressWithPath, ApiByronWallet, ApiT (..) )
 import Cardano.Wallet.Primitive.NetworkId
     ( HasSNetworkId (..) )
 import Cardano.Wallet.Primitive.Types.Address
     ( AddressState (..) )
-import Cardano.Wallet.Shelley.Compatibility
-    ( encodeAddress )
 import Control.Monad
     ( forM_ )
 import Control.Monad.Trans.Resource
